refactor: parallelize `object_integration_test`
Split the longer-running tests of `object_integration_test.cc` into
their own files, so we can run them in parallel.

The longest component test executes for only half the time of what
Carlos reported (115s vs 230s), although my timings were generally
slower than his to begin with. Also note #3872 has a drastic effect
on the two longest running tests, if we address that the runtime
should drop to ~47 seconds which is a 5x speedup.

Part of #3854
