+++ google/cloud/storage/tests/CMakeLists.txt
@@ -23,13 +23,20 @@ set(storage_client_integration_tests
     curl_sign_blob_integration_test.cc
     error_injection_integration_test.cc
     key_file_integration_test.cc
+    object_basic_crud_integration_test.cc
     object_checksum_integration_test.cc
+    object_compose_many_integration_test.cc
     object_file_integration_test.cc
     object_file_multi_threaded_test.cc
     object_hash_integration_test.cc
     object_insert_integration_test.cc
     object_integration_test.cc
+    object_list_objects_versions_integration_test.cc
     object_media_integration_test.cc
+    object_parallel_upload_integration_test.cc
+    object_plenty_clients_serially_integration_test.cc
+    object_plenty_clients_simultaneously_integration_test.cc
+    object_resumable_parallel_upload_integration_test.cc
     object_resumable_write_integration_test.cc
     object_rewrite_integration_test.cc
     object_write_streambuf_integration_test.cc
