+++ google/cloud/storage/tests/object_basic_crud_integration_test.cc
@@ -1,0 +1,154 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/storage/client.h"
+#include "google/cloud/storage/testing/object_integration_test.h"
+#include "google/cloud/storage/testing/storage_integration_test.h"
+#include "google/cloud/log.h"
+#include "google/cloud/status.h"
+#include "google/cloud/status_or.h"
+#include "google/cloud/testing_util/assert_ok.h"
+#include "google/cloud/testing_util/expect_exception.h"
+#include <gmock/gmock.h>
+#include <sys/types.h>
+#include <algorithm>
+#include <cstring>
+#include <iostream>
+#include <memory>
+#include <string>
+
+namespace google {
+namespace cloud {
+namespace storage {
+inline namespace STORAGE_CLIENT_NS {
+namespace {
+
+using ObjectBasicCRUDIntegrationTest =
+    ::google::cloud::storage::testing::ObjectIntegrationTest;
+
+/// @test Verify the Object CRUD (Create, Get, Update, Delete, List) operations.
+TEST_F(ObjectBasicCRUDIntegrationTest, BasicCRUD) {
+  StatusOr<Client> client = MakeIntegrationTestClient();
+  ASSERT_STATUS_OK(client);
+
+  auto objects = client->ListObjects(bucket_name_);
+  std::vector<ObjectMetadata> initial_list;
+  for (auto&& o : objects) {
+    initial_list.emplace_back(std::move(o).value());
+  }
+
+  auto name_counter = [](std::string const& name,
+                         std::vector<ObjectMetadata> const& list) {
+    return std::count_if(
+        list.begin(), list.end(),
+        [name](ObjectMetadata const& m) { return m.name() == name; });
+  };
+
+  auto object_name = MakeRandomObjectName();
+  ASSERT_EQ(0, name_counter(object_name, initial_list))
+      << "Test aborted. The object <" << object_name << "> already exists."
+      << "This is unexpected as the test generates a random object name.";
+
+  // Create the object, but only if it does not exist already.
+  StatusOr<ObjectMetadata> insert_meta =
+      client->InsertObject(bucket_name_, object_name, LoremIpsum(),
+                           IfGenerationMatch(0), Projection("full"));
+  ASSERT_STATUS_OK(insert_meta);
+
+  objects = client->ListObjects(bucket_name_);
+
+  std::vector<ObjectMetadata> current_list;
+  for (auto&& o : objects) {
+    current_list.emplace_back(std::move(o).value());
+  }
+  EXPECT_EQ(1, name_counter(object_name, current_list));
+
+  StatusOr<ObjectMetadata> get_meta = client->GetObjectMetadata(
+      bucket_name_, object_name, Generation(insert_meta->generation()),
+      Projection("full"));
+  ASSERT_STATUS_OK(get_meta);
+  EXPECT_EQ(*get_meta, *insert_meta);
+
+  ObjectMetadata update = *get_meta;
+  update.mutable_acl().emplace_back(
+      ObjectAccessControl().set_role("READER").set_entity(
+          "allAuthenticatedUsers"));
+  update.set_cache_control("no-cache")
+      .set_content_disposition("inline")
+      .set_content_encoding("identity")
+      .set_content_language("en")
+      .set_content_type("plain/text");
+  update.mutable_metadata().emplace("updated", "true");
+  StatusOr<ObjectMetadata> updated_meta = client->UpdateObject(
+      bucket_name_, object_name, update, Projection("full"));
+  ASSERT_STATUS_OK(updated_meta);
+
+  // Because some of the ACL values are not predictable we convert the values we
+  // care about to strings and compare that.
+  {
+    auto acl_to_string_vector =
+        [](std::vector<ObjectAccessControl> const& acl) {
+          std::vector<std::string> v;
+          std::transform(acl.begin(), acl.end(), std::back_inserter(v),
+                         [](ObjectAccessControl const& x) {
+                           return x.entity() + " = " + x.role();
+                         });
+          return v;
+        };
+    auto expected = acl_to_string_vector(update.acl());
+    auto actual = acl_to_string_vector(updated_meta->acl());
+    EXPECT_THAT(expected, ::testing::UnorderedElementsAreArray(actual));
+  }
+  EXPECT_EQ(update.cache_control(), updated_meta->cache_control())
+      << *updated_meta;
+  EXPECT_EQ(update.content_disposition(), updated_meta->content_disposition())
+      << *updated_meta;
+  EXPECT_EQ(update.content_encoding(), updated_meta->content_encoding())
+      << *updated_meta;
+  EXPECT_EQ(update.content_language(), updated_meta->content_language())
+      << *updated_meta;
+  EXPECT_EQ(update.content_type(), updated_meta->content_type())
+      << *updated_meta;
+  EXPECT_EQ(update.metadata(), updated_meta->metadata()) << *updated_meta;
+
+  ObjectMetadata desired_patch = *updated_meta;
+  desired_patch.set_content_language("en");
+  desired_patch.mutable_metadata().erase("updated");
+  desired_patch.mutable_metadata().emplace("patched", "true");
+  StatusOr<ObjectMetadata> patched_meta = client->PatchObject(
+      bucket_name_, object_name, *updated_meta, desired_patch);
+  ASSERT_STATUS_OK(patched_meta);
+
+  EXPECT_EQ(desired_patch.metadata(), patched_meta->metadata())
+      << *patched_meta;
+  EXPECT_EQ(desired_patch.content_language(), patched_meta->content_language())
+      << *patched_meta;
+
+  auto status = client->DeleteObject(bucket_name_, object_name);
+  ASSERT_STATUS_OK(status);
+
+  objects = client->ListObjects(bucket_name_);
+  current_list.clear();
+  for (auto&& o : objects) {
+    current_list.emplace_back(std::move(o).value());
+  }
+
+  EXPECT_EQ(0, name_counter(object_name, current_list));
+}
+
+}  // anonymous namespace
+}  // namespace STORAGE_CLIENT_NS
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
