+++ google/cloud/storage/tests/object_integration_test.cc
@@ -12,25 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "google/cloud/storage/testing/object_integration_test.h"
 #include "google/cloud/storage/client.h"
-#include "google/cloud/storage/parallel_upload.h"
 #include "google/cloud/storage/testing/storage_integration_test.h"
-#include "google/cloud/storage/testing/temp_file.h"
-#include "google/cloud/internal/getenv.h"
 #include "google/cloud/log.h"
+#include "google/cloud/status.h"
+#include "google/cloud/status_or.h"
 #include "google/cloud/testing_util/assert_ok.h"
-#include "google/cloud/testing_util/capture_log_lines_backend.h"
 #include "google/cloud/testing_util/expect_exception.h"
 #include <gmock/gmock.h>
 #include <sys/types.h>
-#include <cstddef>
+#include <algorithm>
 #include <cstring>
-#ifdef __unix__
-#include <dirent.h>
-#endif  // __unix__
 #include <iostream>
 #include <memory>
-#include <regex>
 #include <string>
 
 namespace google {
@@ -39,142 +34,11 @@ namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace {
 
-using ::google::cloud::storage::testing::CountMatchingEntities;
 using ::google::cloud::storage::testing::TestPermanentFailure;
-using ::testing::HasSubstr;
 using ::testing::UnorderedElementsAre;
 
-class ObjectIntegrationTest
-    : public google::cloud::storage::testing::StorageIntegrationTest {
- protected:
-  void SetUp() override {
-    project_id_ =
-        google::cloud::internal::GetEnv("GOOGLE_CLOUD_PROJECT").value_or("");
-    ASSERT_FALSE(project_id_.empty());
-    bucket_name_ = google::cloud::internal::GetEnv(
-                       "GOOGLE_CLOUD_CPP_STORAGE_TEST_BUCKET_NAME")
-                       .value_or("");
-    ASSERT_FALSE(bucket_name_.empty());
-  }
-
-  std::string MakeEntityName() {
-    // We always use the viewers for the project because it is known to exist.
-    return "project-viewers-" + project_id_;
-  }
-
-  std::string project_id_;
-  std::string bucket_name_;
-};
-
-/// @test Verify the Object CRUD (Create, Get, Update, Delete, List) operations.
-TEST_F(ObjectIntegrationTest, BasicCRUD) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto objects = client->ListObjects(bucket_name_);
-  std::vector<ObjectMetadata> initial_list;
-  for (auto&& o : objects) {
-    initial_list.emplace_back(std::move(o).value());
-  }
-
-  auto name_counter = [](std::string const& name,
-                         std::vector<ObjectMetadata> const& list) {
-    return std::count_if(
-        list.begin(), list.end(),
-        [name](ObjectMetadata const& m) { return m.name() == name; });
-  };
-
-  auto object_name = MakeRandomObjectName();
-  ASSERT_EQ(0, name_counter(object_name, initial_list))
-      << "Test aborted. The object <" << object_name << "> already exists."
-      << "This is unexpected as the test generates a random object name.";
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> insert_meta =
-      client->InsertObject(bucket_name_, object_name, LoremIpsum(),
-                           IfGenerationMatch(0), Projection("full"));
-  ASSERT_STATUS_OK(insert_meta);
-
-  objects = client->ListObjects(bucket_name_);
-
-  std::vector<ObjectMetadata> current_list;
-  for (auto&& o : objects) {
-    current_list.emplace_back(std::move(o).value());
-  }
-  EXPECT_EQ(1, name_counter(object_name, current_list));
-
-  StatusOr<ObjectMetadata> get_meta = client->GetObjectMetadata(
-      bucket_name_, object_name, Generation(insert_meta->generation()),
-      Projection("full"));
-  ASSERT_STATUS_OK(get_meta);
-  EXPECT_EQ(*get_meta, *insert_meta);
-
-  ObjectMetadata update = *get_meta;
-  update.mutable_acl().emplace_back(
-      ObjectAccessControl().set_role("READER").set_entity(
-          "allAuthenticatedUsers"));
-  update.set_cache_control("no-cache")
-      .set_content_disposition("inline")
-      .set_content_encoding("identity")
-      .set_content_language("en")
-      .set_content_type("plain/text");
-  update.mutable_metadata().emplace("updated", "true");
-  StatusOr<ObjectMetadata> updated_meta = client->UpdateObject(
-      bucket_name_, object_name, update, Projection("full"));
-  ASSERT_STATUS_OK(updated_meta);
-
-  // Because some of the ACL values are not predictable we convert the values we
-  // care about to strings and compare that.
-  {
-    auto acl_to_string_vector =
-        [](std::vector<ObjectAccessControl> const& acl) {
-          std::vector<std::string> v;
-          std::transform(acl.begin(), acl.end(), std::back_inserter(v),
-                         [](ObjectAccessControl const& x) {
-                           return x.entity() + " = " + x.role();
-                         });
-          return v;
-        };
-    auto expected = acl_to_string_vector(update.acl());
-    auto actual = acl_to_string_vector(updated_meta->acl());
-    EXPECT_THAT(expected, ::testing::UnorderedElementsAreArray(actual));
-  }
-  EXPECT_EQ(update.cache_control(), updated_meta->cache_control())
-      << *updated_meta;
-  EXPECT_EQ(update.content_disposition(), updated_meta->content_disposition())
-      << *updated_meta;
-  EXPECT_EQ(update.content_encoding(), updated_meta->content_encoding())
-      << *updated_meta;
-  EXPECT_EQ(update.content_language(), updated_meta->content_language())
-      << *updated_meta;
-  EXPECT_EQ(update.content_type(), updated_meta->content_type())
-      << *updated_meta;
-  EXPECT_EQ(update.metadata(), updated_meta->metadata()) << *updated_meta;
-
-  ObjectMetadata desired_patch = *updated_meta;
-  desired_patch.set_content_language("en");
-  desired_patch.mutable_metadata().erase("updated");
-  desired_patch.mutable_metadata().emplace("patched", "true");
-  StatusOr<ObjectMetadata> patched_meta = client->PatchObject(
-      bucket_name_, object_name, *updated_meta, desired_patch);
-  ASSERT_STATUS_OK(patched_meta);
-
-  EXPECT_EQ(desired_patch.metadata(), patched_meta->metadata())
-      << *patched_meta;
-  EXPECT_EQ(desired_patch.content_language(), patched_meta->content_language())
-      << *patched_meta;
-
-  auto status = client->DeleteObject(bucket_name_, object_name);
-  ASSERT_STATUS_OK(status);
-
-  objects = client->ListObjects(bucket_name_);
-  current_list.clear();
-  for (auto&& o : objects) {
-    current_list.emplace_back(std::move(o).value());
-  }
-
-  EXPECT_EQ(0, name_counter(object_name, current_list));
-}
+using ObjectIntegrationTest =
+    ::google::cloud::storage::testing::ObjectIntegrationTest;
 
 TEST_F(ObjectIntegrationTest, FullPatch) {
   StatusOr<Client> client = MakeIntegrationTestClient();
@@ -249,77 +113,6 @@ TEST_F(ObjectIntegrationTest, FullPatch) {
   ASSERT_STATUS_OK(status);
 }
 
-TEST_F(ObjectIntegrationTest, ListObjectsVersions) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  // This test requires the bucket to be configured with versioning. The buckets
-  // used by the CI build are already configured with versioning enabled. The
-  // bucket created in the testbench also has versioning. Regardless, set the
-  // bucket to the desired state, which will produce a better error message if
-  // there is a configuration problem.
-  auto bucket_meta = client->GetBucketMetadata(bucket_name_);
-  ASSERT_STATUS_OK(bucket_meta);
-  auto updated_meta = client->PatchBucket(
-      bucket_name_,
-      BucketMetadataPatchBuilder().SetVersioning(BucketVersioning{true}),
-      IfMetagenerationMatch(bucket_meta->metageneration()));
-  ASSERT_STATUS_OK(updated_meta);
-  ASSERT_TRUE(updated_meta->versioning().has_value());
-  ASSERT_TRUE(updated_meta->versioning().value().enabled);
-
-  auto create_object_with_3_versions = [&client, this] {
-    auto object_name = MakeRandomObjectName();
-    // ASSERT_TRUE does not work inside this lambda because the return type is
-    // not `void`, use `.value()` instead to throw (or crash) on the unexpected
-    // error.
-    auto meta = client
-                    ->InsertObject(bucket_name_, object_name,
-                                   "contents for the first revision",
-                                   storage::IfGenerationMatch(0))
-                    .value();
-    client
-        ->InsertObject(bucket_name_, object_name,
-                       "contents for the second revision")
-        .value();
-    client
-        ->InsertObject(bucket_name_, object_name,
-                       "contents for the final revision")
-        .value();
-    return meta.name();
-  };
-
-  std::vector<std::string> expected(4);
-  std::generate_n(expected.begin(), expected.size(),
-                  create_object_with_3_versions);
-
-  ListObjectsReader reader = client->ListObjects(bucket_name_, Versions(true));
-  std::vector<std::string> actual;
-  for (auto it = reader.begin(); it != reader.end(); ++it) {
-    auto const& meta = it->value();
-    EXPECT_EQ(bucket_name_, meta.bucket());
-    actual.push_back(meta.name());
-  }
-  auto produce_joined_list = [&actual] {
-    std::string joined;
-    for (auto const& x : actual) {
-      joined += "  ";
-      joined += x;
-      joined += "\n";
-    }
-    return joined;
-  };
-  // There may be a lot of other objects in the bucket, so we want to verify
-  // that any objects we created are found there, but cannot expect a perfect
-  // match.
-
-  for (auto const& name : expected) {
-    EXPECT_EQ(3, std::count(actual.begin(), actual.end(), name))
-        << "Expected to find 3 copies of " << name << " in the object list:\n"
-        << produce_joined_list();
-  }
-}
-
 TEST_F(ObjectIntegrationTest, ListObjectsDelimiter) {
   if (UsingTestbench()) GTEST_SKIP();
 
@@ -385,145 +178,6 @@ TEST_F(ObjectIntegrationTest, BasicReadWrite) {
   auto stream = client->ReadObject(bucket_name_, object_name);
   std::string actual(std::istreambuf_iterator<char>{stream}, {});
   EXPECT_EQ(expected, actual);
-
-  auto status = client->DeleteObject(bucket_name_, object_name);
-  ASSERT_STATUS_OK(status);
-}
-
-StatusOr<std::size_t> GetNumEntries(std::string const& path) {
-#ifdef __unix__
-  dirent** namelist;
-  int res = scandir(path.c_str(), &namelist, nullptr, alphasort);
-  if (res < 0) {
-    return Status(StatusCode::kInternal, "Failed to open directory \"" + path +
-                                             "\": " + strerror(errno));
-  }
-  for (int i = 0; i < res; ++i) {
-    free(namelist[i]);
-  }
-  free(namelist);
-  return res;
-#else   // __unix__
-  return Status(StatusCode::kUnimplemented,
-                "Can't check #entries in " + path +
-                    ", because only UNIX systems are supported");
-#endif  // __unix__
-}
-
-StatusOr<std::size_t> GetNumOpenFiles() {
-  auto res = GetNumEntries("/proc/self/fd");
-  if (!res) {
-    return res;
-  }
-  if (*res < 3) {
-    return Status(StatusCode::kInternal,
-                  "Expected at least three entries in /proc/self/fd: ., .., "
-                  "and the directory itself, found " +
-                      std::to_string(*res));
-  }
-  return *res - 3;
-}
-
-TEST_F(ObjectIntegrationTest, PlentyClientsSimultaneously) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto object_name = MakeRandomObjectName();
-
-  std::string expected = LoremIpsum();
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> meta = client->InsertObject(
-      bucket_name_, object_name, expected, IfGenerationMatch(0));
-  ASSERT_STATUS_OK(meta);
-
-  EXPECT_EQ(object_name, meta->name());
-  EXPECT_EQ(bucket_name_, meta->bucket());
-
-  // Create a iostream to read the object back.
-  auto num_fds_before_test = GetNumOpenFiles();
-  std::vector<Client> read_clients;
-  std::vector<ObjectReadStream> read_streams;
-  for (int i = 0; i != 100; ++i) {
-    auto read_client = MakeIntegrationTestClient();
-    ASSERT_STATUS_OK(read_client);
-    auto stream = read_client->ReadObject(bucket_name_, object_name);
-    char c;
-    stream.read(&c, 1);
-    read_streams.emplace_back(std::move(stream));
-    read_clients.emplace_back(*std::move(read_client));
-  }
-  auto num_fds_during_test = GetNumOpenFiles();
-  read_streams.clear();
-  read_clients.clear();
-  auto num_fds_after_test = GetNumOpenFiles();
-
-#ifdef __linux__
-  ASSERT_STATUS_OK(num_fds_before_test);
-  ASSERT_STATUS_OK(num_fds_during_test);
-  ASSERT_STATUS_OK(num_fds_after_test);
-
-  EXPECT_LT(*num_fds_before_test, *num_fds_during_test)
-      << "Clients keeps at least some file descriptors open";
-  EXPECT_LT(*num_fds_after_test, *num_fds_during_test)
-      << "Releasing clients also releases at least some file descriptors";
-  EXPECT_EQ(*num_fds_before_test, *num_fds_after_test)
-      << "Clients are leaking descriptors";
-#endif  // __linux__
-
-  auto status = client->DeleteObject(bucket_name_, object_name);
-  ASSERT_STATUS_OK(status);
-}
-
-TEST_F(ObjectIntegrationTest, PlentyClientsSerially) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto object_name = MakeRandomObjectName();
-
-  std::string expected = LoremIpsum();
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> meta = client->InsertObject(
-      bucket_name_, object_name, expected, IfGenerationMatch(0));
-  ASSERT_STATUS_OK(meta);
-
-  EXPECT_EQ(object_name, meta->name());
-  EXPECT_EQ(bucket_name_, meta->bucket());
-
-  // Create a iostream to read the object back.
-  auto num_fds_before_test = GetNumOpenFiles();
-#ifdef __linux__
-  std::size_t delta = 0;
-#endif  // __linux__
-  for (int i = 0; i != 100; ++i) {
-    auto read_client = MakeIntegrationTestClient();
-    ASSERT_STATUS_OK(read_client);
-    auto stream = read_client->ReadObject(bucket_name_, object_name);
-    char c;
-    stream.read(&c, 1);
-#ifdef __linux__
-    auto num_fds_during_test = GetNumOpenFiles();
-    ASSERT_STATUS_OK(num_fds_before_test);
-    ASSERT_STATUS_OK(num_fds_during_test);
-    if (delta == 0) {
-      delta = *num_fds_during_test - *num_fds_before_test;
-    }
-    EXPECT_GE(*num_fds_before_test + delta, *num_fds_during_test)
-        << "Expect each client to create the same number of file descriptors"
-        << ", num_fds_before_test=" << *num_fds_before_test
-        << ", num_fds_during_test=" << *num_fds_during_test
-        << ", delta=" << delta;
-#endif  // __linux__
-  }
-#ifdef __linux__
-  auto num_fds_after_test = GetNumOpenFiles();
-  ASSERT_STATUS_OK(num_fds_before_test);
-  ASSERT_STATUS_OK(num_fds_after_test);
-
-  EXPECT_EQ(*num_fds_before_test, *num_fds_after_test)
-      << "Clients are leaking descriptors";
-#endif  // __linux__
 
   auto status = client->DeleteObject(bucket_name_, object_name);
   ASSERT_STATUS_OK(status);
@@ -1012,171 +666,6 @@ TEST_F(ObjectIntegrationTest, DeleteAccessControlFailure) {
   // This operation should fail because the source object does not exist.
   auto status = client->DeleteObjectAcl(bucket_name_, object_name, entity_name);
   ASSERT_FALSE(status.ok());
-}
-
-TEST_F(ObjectIntegrationTest, ComposeMany) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto prefix = CreateRandomPrefixName();
-  std::string const dest_object_name = prefix + ".dest";
-
-  std::vector<ComposeSourceObject> source_objs;
-  std::string expected;
-  for (int i = 0; i != 100; ++i) {
-    std::string const object_name = prefix + ".src-" + std::to_string(i);
-    std::string content = std::to_string(i);
-    expected += content;
-    StatusOr<ObjectMetadata> insert_meta = client->InsertObject(
-        bucket_name_, object_name, std::move(content), IfGenerationMatch(0));
-    ASSERT_STATUS_OK(insert_meta);
-    source_objs.emplace_back(ComposeSourceObject{
-        std::move(object_name), insert_meta->generation(), {}});
-  }
-
-  auto res = ComposeMany(*client, bucket_name_, std::move(source_objs), prefix,
-                         dest_object_name, false);
-
-  ASSERT_STATUS_OK(res);
-  EXPECT_EQ(dest_object_name, res->name());
-
-  auto stream = client->ReadObject(bucket_name_, dest_object_name);
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  EXPECT_EQ(expected, actual);
-
-  auto deletion_status = client->DeleteObject(
-      bucket_name_, dest_object_name, IfGenerationMatch(res->generation()));
-  ASSERT_STATUS_OK(deletion_status);
-}
-
-TEST_F(ObjectIntegrationTest, ParallelUpload) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto prefix = CreateRandomPrefixName();
-  std::string const dest_object_name = prefix + ".dest";
-
-  testing::TempFile temp_file(LoremIpsum());
-
-  auto object_metadata = ParallelUploadFile(
-      *client, temp_file.name(), bucket_name_, dest_object_name, prefix, false,
-      MinStreamSize(0), IfGenerationMatch(0));
-
-  auto stream =
-      client->ReadObject(bucket_name_, dest_object_name,
-                         IfGenerationMatch(object_metadata->generation()));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  EXPECT_EQ(LoremIpsum(), actual);
-
-  std::vector<ObjectMetadata> objects;
-  for (auto& object : client->ListObjects(bucket_name_)) {
-    ASSERT_STATUS_OK(object);
-    if (object->name().substr(0, prefix.size()) == prefix &&
-        object->name() != prefix) {
-      objects.emplace_back(*std::move(object));
-    }
-  }
-  ASSERT_EQ(1U, objects.size());
-  ASSERT_EQ(prefix + ".dest", objects[0].name());
-
-  auto deletion_status =
-      client->DeleteObject(bucket_name_, dest_object_name,
-                           IfGenerationMatch(object_metadata->generation()));
-  ASSERT_STATUS_OK(deletion_status);
-}
-
-TEST_F(ObjectIntegrationTest, ResumableParallelUpload) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto prefix = CreateRandomPrefixName();
-  std::string const dest_object_name = prefix + ".dest";
-
-  std::string resumable_session_id;
-  {
-    auto state =
-        PrepareParallelUpload(*client, bucket_name_, dest_object_name, 2,
-                              prefix, UseResumableUploadSession(""));
-    ASSERT_STATUS_OK(state);
-    resumable_session_id = state->resumable_session_id();
-    state->shards()[0] << "1";
-    std::move(state->shards()[0]).Suspend();
-    state->shards()[1] << "34";
-    state->shards().clear();
-  }
-  std::int64_t res_gen;
-  {
-    auto state = PrepareParallelUpload(
-        *client, bucket_name_, dest_object_name, 2, prefix,
-        UseResumableUploadSession(resumable_session_id));
-    ASSERT_STATUS_OK(state);
-    ASSERT_EQ(0, state->shards()[0].next_expected_byte());
-    state->shards()[0] << "12";
-    state->shards().clear();
-    auto res = state->WaitForCompletion().get();
-    ASSERT_STATUS_OK(res);
-    res_gen = res->generation();
-  }
-  auto stream = client->ReadObject(bucket_name_, dest_object_name,
-                                   IfGenerationMatch(res_gen));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  EXPECT_EQ("1234", actual);
-  auto deletion_status = client->DeleteObject(bucket_name_, dest_object_name,
-                                              IfGenerationMatch(res_gen));
-  ASSERT_STATUS_OK(deletion_status);
-}
-
-TEST_F(ObjectIntegrationTest, ResumeParallelUploadFile) {
-  StatusOr<Client> client = MakeIntegrationTestClient();
-  ASSERT_STATUS_OK(client);
-
-  auto prefix = CreateRandomPrefixName();
-  std::string const dest_object_name = prefix + ".dest";
-
-  testing::TempFile temp_file(LoremIpsum());
-
-  auto shards = CreateUploadShards(*client, temp_file.name(), bucket_name_,
-                                   dest_object_name, prefix, MinStreamSize(0),
-                                   MaxStreams(3), IfGenerationMatch(0),
-                                   UseResumableUploadSession(""));
-
-  ASSERT_STATUS_OK(shards);
-  ASSERT_GT(shards->size(), 1);
-
-  std::string resumable_session_id = (*shards)[0].resumable_session_id();
-  auto first_part_res_future = (*shards)[0].WaitForCompletion();
-  ASSERT_STATUS_OK((*shards)[0].Upload());
-  shards->clear();  // we'll resume those
-  auto first_part_res = first_part_res_future.get();
-  EXPECT_FALSE(first_part_res);
-  EXPECT_EQ(StatusCode::kCancelled, first_part_res.status().code());
-
-  auto object_metadata = ParallelUploadFile(
-      *client, temp_file.name(), bucket_name_, dest_object_name, prefix, false,
-      MinStreamSize(0), IfGenerationMatch(0),
-      UseResumableUploadSession(resumable_session_id));
-
-  auto stream =
-      client->ReadObject(bucket_name_, dest_object_name,
-                         IfGenerationMatch(object_metadata->generation()));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  EXPECT_EQ(LoremIpsum(), actual);
-
-  std::vector<ObjectMetadata> objects;
-  for (auto& object : client->ListObjects(bucket_name_)) {
-    ASSERT_STATUS_OK(object);
-    if (object->name().substr(0, prefix.size()) == prefix &&
-        object->name() != prefix) {
-      objects.emplace_back(*std::move(object));
-    }
-  }
-  ASSERT_EQ(1U, objects.size());
-  ASSERT_EQ(prefix + ".dest", objects[0].name());
-
-  auto deletion_status =
-      client->DeleteObject(bucket_name_, dest_object_name,
-                           IfGenerationMatch(object_metadata->generation()));
-  ASSERT_STATUS_OK(deletion_status);
 }
 
 }  // anonymous namespace
